+++ src/src/Course.java
@@ -18,6 +18,12 @@ public class Course {
         gradebook = new Gradebook();
         roster = new Vector<Student>();
     }
+    public Course(Gradebook gb, Vector<Student> rost){
+    
+        gradebook = gb;
+        roster = new Vector<Student>(rost);
+    
+    }
 
     // Functions
     public void calculateOverallGrade(int StudentID){
@@ -73,169 +79,3 @@ public class Course {
     }
 };
 
-// ---- Data Classes ----
-
-// Student
-class Student {
-    private String firstName;
-    private String lastName;
-    private float overallGrade;
-    private int ID;
-    private Vector<Float> scores;
-
-    // Constructor
-    public Student(){
-        firstName = "Undefined firstName";
-        lastName = "Undefined lastName";
-        overallGrade = -4;
-        ID = -5;
-        scores = new Vector<Float>();
-    }
-
-    // Get/Set
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public float getOverallGrade() {
-        return overallGrade;
-    }
-
-    public void setOverallGrade(float overallGrade) {
-        this.overallGrade = overallGrade;
-    }
-
-    public int getID() {
-        return ID;
-    }
-
-    public void setID(int ID) {
-        this.ID = ID;
-    }
-
-    public Vector<Float> getScores() {
-        return scores;
-    }
-
-    public void setScores(Vector<Float> scores) {
-        this.scores = scores;
-    }
-
-    public void setScore(int AssignmentID, float score){
-        this.scores.set(AssignmentID, score);
-    }
-};
-
-// Assignment
-class Assignment {
-    private String name;
-    // Category that the assignment belongs to based on numerical ID of category
-    private int category;
-    private int maxScore;
-    
-
-    // Constructor
-    public Assignment(){
-        name = "Undefined Assignment Name";
-        category = -6;
-        maxScore = -7;
-    }
-
-    // Get/Set
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getCategory() {
-        return category;
-    }
-
-    public void setCategory(int category) {
-        this.category = category;
-    }
-
-    public int getMaxScore() {
-        return maxScore;
-    }
-
-    public void setMaxScore(int maxScore) {
-        this.maxScore = maxScore;
-    }
-};
-
-// Gradebook
-class Gradebook {
-    private Vector<Assignment> assignments;
-    // Scale works by integers setting cutoff per grade (I.E, if percent score is below scale[0], it is an F; if percent score is not below score[3], it must be an A)
-    // Default scale would be scale = [60,70,80,90]
-    private int[] scale;
-    private Vector<String> catNames;
-    private Vector<Integer> catID;
-    private Vector<Float> catWeights;
-
-    // Constructor
-    public Gradebook() {
-        assignments = new Vector<Assignment>();
-        scale = new int[]{60, 70, 80, 90};
-        catNames = new Vector<String>();
-        catID = new Vector<Integer>();
-        catWeights = new Vector<Float>();
-    }
-
-    // Get/Set
-    public Vector<Assignment> getAssignments() {
-        return assignments;
-    }
-
-    public void setAssignments(Vector<Assignment> assignments) {
-        this.assignments = assignments;
-    }
-
-    public int[] getScale() {
-        return scale;
-    }
-
-    public void setScale(int[] scale) {
-        this.scale = scale;
-    }
-
-    public Vector<String> getCatNames() {
-        return catNames;
-    }
-
-    public void setCatNames(Vector<String> catNames) {
-        this.catNames = catNames;
-    }
-
-    public Vector<Integer> getCatID() {
-        return catID;
-    }
-
-    public void setCatID(Vector<Integer> catID) {
-        this.catID = catID;
-    }
-
-    public Vector<Float> getCatWeights() {
-        return catWeights;
-    }
-
-    public void setCatWeights(Vector<Float> catWeights) {
-        this.catWeights = catWeights;
-    }
-};
